Order jobs by_priority more like original gem
The original delayed_job_sequel gem, when asked to order by_priority,
orders by both the priority and run_at columns. We removed the priority
sort in an earlier commit on this fork, but would like it to be
reinstated so we can usefully control the priority of errored jobs.

[#109807570]

Signed-off-by: Dan Wendorf <dwendorf@pivotal.io>